+++ kscan/src/commonMain/kotlin/org/ncgroup/kscan/view/ScannerViewImpl.kt
@@ -17,6 +17,5 @@ internal expect fun ScannerViewImpl(
     cameraId: String?,
     scannerController: ScannerController?,
     filter: (Barcode) -> Boolean,
-    autoZoom: Boolean,
     result: (BarcodeResult) -> Unit,
 )