+++ kscan/src/jvmMain/kotlin/org/ncgroup/kscan/view/ScannerViewImpl.jvm.kt
@@ -43,7 +43,6 @@ internal actual fun ScannerViewImpl(
     cameraId: String?,
     scannerController: ScannerController?,
     filter: (Barcode) -> Boolean,
-    autoZoom: Boolean,
     result: (BarcodeResult) -> Unit,
 ) {
     val updatedResult by rememberUpdatedState(result)